Moved verifier into separate verifier package
